LayoutTests/http/wpt/service-workers/claim-worker-fetch-without-nesting.https-expected.txt


PASS fetch() in Worker should be intercepted after the client is claimed.
PASS fetch() in blob URL Worker should be intercepted after the client is claimed.
